Create seperate task to deploy external resources
