+++ azure-webapp-maven-plugin/src/main/java/com/microsoft/azure/maven/webapp/DeployMojo.java
@@ -7,11 +7,8 @@ package com.microsoft.azure.maven.webapp;
 
 import com.microsoft.azure.maven.model.DeploymentResource;
 import com.microsoft.azure.maven.webapp.configuration.DeploymentSlotConfig;
-import com.microsoft.azure.maven.webapp.utils.FTPUtils;
-import com.microsoft.azure.maven.webapp.utils.Utils;
-import com.microsoft.azure.toolkit.lib.appservice.model.PublishingProfile;
+import com.microsoft.azure.maven.webapp.task.DeployExternalResourcesTask;
 import com.microsoft.azure.toolkit.lib.appservice.model.WebAppArtifact;
-import com.microsoft.azure.toolkit.lib.appservice.service.IAppService;
 import com.microsoft.azure.toolkit.lib.appservice.service.IWebApp;
 import com.microsoft.azure.toolkit.lib.appservice.service.IWebAppBase;
 import com.microsoft.azure.toolkit.lib.appservice.service.IWebAppDeploymentSlot;
@@ -20,12 +17,9 @@ import com.microsoft.azure.toolkit.lib.appservice.task.DeployWebAppTask;
 import com.microsoft.azure.toolkit.lib.common.bundle.AzureString;
 import com.microsoft.azure.toolkit.lib.common.exception.AzureExecutionException;
 import com.microsoft.azure.toolkit.lib.common.messager.AzureMessager;
-import org.apache.commons.net.ftp.FTPClient;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 
-import java.io.File;
-import java.io.IOException;
 import java.util.List;
 
 /**
@@ -88,28 +82,7 @@ public class DeployMojo extends AbstractWebAppMojo {
         new DeployWebAppTask(target, artifacts, isStopAppDuringDeployment()).execute();
     }
 
-    private void deployExternalResources(final IAppService<?> target, final List<DeploymentResource> resources) throws AzureExecutionException {
-        if (resources.isEmpty()) {
-            return;
-        }
-        AzureMessager.getMessager().info(AzureString.format("Uploading resources to %s", target.name()));
-        final PublishingProfile publishingProfile = target.getPublishingProfile();
-        final String serverUrl = publishingProfile.getFtpUrl().split("/", 2)[0];
-        try {
-            final FTPClient ftpClient = FTPUtils.getFTPClient(serverUrl, publishingProfile.getFtpUsername(), publishingProfile.getFtpPassword());
-            for (final DeploymentResource externalResource : resources) {
-                uploadResource(externalResource, ftpClient);
-            }
-        } catch (IOException e) {
-            throw new AzureExecutionException(e.getMessage(), e);
-        }
-    }
-
-    private static void uploadResource(DeploymentResource resource, FTPClient ftpClient) throws IOException {
-        final List<File> files = Utils.getArtifacts(resource);
-        final String target = resource.getAbsoluteTargetPath();
-        for (final File file : files) {
-            FTPUtils.uploadFile(ftpClient, file.getPath(), target);
-        }
+    private void deployExternalResources(final IWebAppBase<?> target, final List<DeploymentResource> resources) {
+        new DeployExternalResourcesTask(target, resources).execute();
     }
 }
